Refactor stage_train and stage_validate and stage data into one file
